% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.cjdiag_tree}
\alias{plot.cjdiag_tree}
\title{Plot Decision Tree}
\usage{
\method{plot}{cjdiag_tree}(x, ...)
}
\arguments{
\item{x}{A \code{cjdiag_tree} object from \code{\link[=cj_fit]{cj_fit()}}}

\item{...}{Additional arguments passed to \code{\link[rpart.plot:rpart.plot]{rpart.plot::rpart.plot()}}}
}
\value{
Invisible NULL (called for side effect of plotting)
}
\description{
Renders the decision tree using \code{\link[rpart.plot:rpart.plot]{rpart.plot::rpart.plot()}}.
}
\examples{
\donttest{
df <- data.frame(
  y = sample(0:1, 200, TRUE),
  a = factor(sample(c("x","y"), 200, TRUE)),
  b = factor(sample(c("p","q","r"), 200, TRUE))
)
tr <- cj_fit(y ~ a + b, data = df, method = "tree")
plot(tr)
}
}
\seealso{
Other plotting: 
\code{\link{plot.cjdiag_crt}()},
\code{\link{plot.cjdiag_forest}()},
\code{\link{plot.cjdiag_importance}()},
\code{\link{plot.cjdiag_nmm}()}
}
\concept{plotting}
